third_party/WebKit/LayoutTests/fast/dom/HTMLScriptElement/adopt-node-crash-expected.txt

CONSOLE MESSAGE: line 17: PASS unless crash.
 
